Skip missing install files

+++ minecraft_server_tools/install_client.py
@@ -40,14 +40,20 @@ def install_extras(source_dir=SERVER_DIR, do_optional=True):
     for install_dir in EXTRA_INSTALL_FOLDERS:
         from_dir = os.path.join(source_dir, install_dir)
         to_dir = os.path.join(MINECRAFT_DIR, install_dir)
-        print(f"\t{install_dir}...")
-        shutil.copytree(from_dir, to_dir, dirs_exist_ok=True)
+        if os.path.exists(from_dir):
+            print(f"\t{install_dir}...")
+            shutil.copytree(from_dir, to_dir, dirs_exist_ok=True)
+        else:
+            print(f"\tskipped: {install_dir}")
 
     for install_file in EXTRA_INSTALL_FILES + (OPTIONAL_INSTALL_FILES if do_optional else []):
         from_path = os.path.join(source_dir, install_file)
         to_path = os.path.join(MINECRAFT_DIR, install_file)
-        print(f"\t{install_file}...")
-        shutil.copy(from_path, to_path)
+        if os.path.exists(from_path):
+            print(f"\t{install_file}...")
+            shutil.copy(from_path, to_path)
+        else:
+            print(f"\tskipped: {install_file}")
 
 
 def ensure_forge_client(source_dir=SERVER_DIR):
